remove jitter

+++ arcdrive.py
@@ -30,11 +30,11 @@ def arcdrive(a_star, radius, leftTurn=1, arc=180, speed=1, forward=1,Kp = 1.8,Ki
 
         if forward == 1 and (Lcurr > Lfinal or Rcurr > Rfinal):
             print("Lcurr: {}\tLfinal: {}\tRcurr: {}\tRfinal: {}".format(Lcurr, Lfinal, Rcurr, Rfinal))
-            a_star.motors(speed*105*forward, speed*100*forward)
+            a_star.motors(int(speed*105*forward), int(speed*100*forward))
             break
         if forward == -1 and (Lcurr < Lfinal or Rcurr < Rfinal):
             print("Lcurr: {}\tLfinal: {}\tRcurr: {}\tRfinal: {}".format(Lcurr, Lfinal, Rcurr, Rfinal))
-            a_star.motors(speed*105*forward, speed*100*forward)
+            a_star.motors(int(speed*105*forward), int(speed*100*forward))
             break
 
         # calculate errors (leaning left)
